set up training pipeline for SRI24-intermediate approach
This commit finally sets up the complete training pipeline for the
approach where the SRI24 atlas gets used as an intermediate space that
images have to be warped *through* in order to be registered to other
images. Training is functional now, finally, but the initial results are
not looking good. A lot more work is needed:
- introduce more loss terms: similarity and anatomy for the displacement v
  (without this, there's not a direct way of learning the case where
   an *atlas* gets warped to a fixed *image*.)
- possibly introduce affine pre-registration, if it turns out to be
needed.